+++ CHANGELOG.md
@@ -22,6 +22,7 @@ i.e. `<Major version>.<Minor version>.<Patch version>`
 - Support customize CDR file name and interval
 - Fix routing block action invalid data #82
 - Unable to delete unengaged routing table #80
+- Support ignore ealry media to preanswer class #91
 
 
 ## [v0.5.9] - 2022-07-26